requirements.txt
easyocr
evaluate
google-cloud-vision
numpy
openpyxl
paddleocr
pandas
pillow
pytesseract
torch
transformers
